Create finetunedata.py
We're going to keep the original `finetune.py` for the time being for reference purposes (will be deleted soon). 

However, this is where we're at on the current iteration of this file. There are still a bunch of changes that need to be committed here. Those changes will be explained in greater depth when there's a chance. 

**Ignore Below** 

^^ Formatting above is to test whether `markdown` formatting works in the commit messages.